cast to int for size_t
git-svn-id: svn://svn.macaulay2.com/Macaulay2/trunk@4896 a8012f80-5f13-0410-8dc0-eb77bd940c0d

+++ M2/Macaulay2/e/gbA.cpp
@@ -2508,7 +2508,7 @@ void gbA::show_mem_usage()
       nmonoms += R->gbvector_n_terms(gb[i]->g.f);
       nmonoms += R->gbvector_n_terms(gb[i]->g.fsyz);
     }
-  printf("\nnumber of (nonminimal) gb elements = %ld\n", gb.size());
+  printf("\nnumber of (nonminimal) gb elements = %d\n", (int)gb.size());
   printf("number of monomials                = %ld\n", nmonoms);
 }
